incorporada camara, keys, dibjua mapas. Algunos errores en el atoi. cambiada la forma de guardar el mata. malloc int**. Algunas funciones de debug, Por hacer: testeos de camara, deporar bugs, movimientos...

+++ lib_basic/src/ftlib/ft_atoi.c
@@ -15,7 +15,7 @@
 int	ft_atoi(const char *str)
 {
 	unsigned int	c;
-	int				sing;
+	int	sing;
 
 	c = 0;
 	sing = 1;
@@ -35,7 +35,7 @@ int	ft_atoi(const char *str)
 		return (-1);
 	if (c > 2147483648 && sing == -1)
 		return (0);
-	return (c * sing);
+	return (((int ) c) * sing);
 }
 
 /*
